+++ components/queueManager/Dashboard/QueueManagerDashboardHeader.tsx
@@ -145,26 +145,6 @@ export const QueueManagerDashboardHeader: FC<QueueManagerDashboardHeaderProps> =
                   </div>
                 </div>
 
-                {/* Second row */}
-                {/* <div className='flex items-center pt-2 space-x-3'>
-                  <FontAwesomeIcon
-                    icon={faUsers}
-                    className='h-5 w-5 py-2 cursor-pointer flex items-center text-gray-500 dark:text-gray-400'
-                    aria-hidden='true'
-                  />
-                  <div className='flex justify-center items-center'>
-                    <p className='text-md font-semibold tracking-tight text-left text-gray-900 dark:text-gray-900 mr-1'>
-                      {t('QueueManager.Queue')}:
-                    </p>
-                    <p className='text-md font-bold leading-6 text-center mr-1 text-gray-900 dark:text-gray-900'>
-                      Assistenza Clienti
-                    </p>
-                    <p className='text-md font-bold leading-6 text-center text-gray-900 dark:text-gray-900'>
-                      500
-                    </p>
-                  </div>
-                </div> */}
-
                 {/* Third row */}
                 <span className='pt-3 text-sm'>{getAlarmDescription(alarmsList, alarmsType)}</span>
               </div>
@@ -179,171 +159,6 @@ export const QueueManagerDashboardHeader: FC<QueueManagerDashboardHeaderProps> =
       </div>
     </>
   )
-
-  const cardGroups = [
-    [
-      <div key='alarms' className='mx-3'>
-        <Dropdown items={dropdownItems} position='left' divider={true} className='pl-3'>
-          <div
-            className={`flex items-center justify-between px-4 mt-1 mb-2 bg-gray-100 rounded-md py-1 ${
-              !isEmpty(alarmsList.list) ? 'bg-red-50' : ''
-            }`}
-          >
-            <div className='flex items-center'>
-              <FontAwesomeIcon
-                icon={faTriangleExclamation}
-                className={`h-6 w-6 pr-6 py-2 cursor-pointer flex items-center ${
-                  !isEmpty(alarmsList.list) ? 'text-red-600' : 'text-gray-500 dark:text-gray-400'
-                }`}
-                aria-hidden='true'
-              />
-              <div className='flex flex-col justify-center'>
-                <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-900'>
-                  {(alarmsList.list && Object.keys(alarmsList.list).length) ?? 0}
-                </p>
-                <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-                  {alarmsList.list && Object.keys(alarmsList.list).length === 1
-                    ? t('QueueManager.Alarm')
-                    : t('QueueManager.Alarms')}
-                </p>
-              </div>
-            </div>
-            <FontAwesomeIcon
-              icon={faChevronDown}
-              className='h-3.5 w-3.5 pl-2 py-2 cursor-pointer flex items-center'
-              aria-hidden='true'
-            />
-          </div>
-        </Dropdown>
-      </div>,
-    ],
-    [
-      <div key='notManaged' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faPhone}
-              className='h-6 w-6 cursor-pointer text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {notManaged.count}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.Not managed customers')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-    [
-      <div key='totalAll' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faPhone}
-              className='h-6 w-6 cursor-pointer text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {totalAll}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.Total calls')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-    [
-      <div key='totalAnswered' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faArrowLeft}
-              className='h-6 w-6 cursor-pointer -rotate-45 text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {totalAnswered}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.Answered calls')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-    [
-      <div key='lostCalls' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faMissed}
-              className='h-6 w-6 cursor-pointer text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {totalFailed}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.Lost calls')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-    [
-      <div key='totalInvalid' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faPhoneSlash}
-              className='h-6 w-6 cursor-pointer text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {totalInvalid}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.Invalid calls')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-    [
-      <div key='inProgress' className='mx-3 pt-1'>
-        <div className='flex items-center w-full'>
-          <div className='h-14 w-14 flex items-center justify-center rounded-md bg-emerald-50'>
-            <FontAwesomeIcon
-              icon={faPause}
-              className='h-6 w-6 cursor-pointer text-emerald-600 dark:text-emerald-600'
-              aria-hidden='true'
-            />
-          </div>
-          <div className='flex flex-col justify-center ml-4'>
-            <p className='text-3xl font-semibold tracking-tight text-left text-gray-900 dark:text-gray-100'>
-              {totalInProgress}
-            </p>
-            <p className='text-sm font-medium leading-6 text-center text-gray-500 dark:text-gray-400'>
-              {t('QueueManager.In progress')}
-            </p>
-          </div>
-        </div>
-      </div>,
-    ],
-  ]
 
   return (
     <>
